+++ examples/from_edn.rs
@@ -3,7 +3,7 @@ use edn_rs::{map, Deserialize, Edn, EdnError, Map};
 #[derive(Debug, PartialEq)]
 struct Person {
     name: String,
-    age: usize,
+    age: u64,
 }
 
 impl Deserialize for Person {
